+++ idea-p4server/impl/src/test/java/net/groboclown/p4/server/impl/connection/FailP4RequestErrorHandler.java
@@ -42,6 +42,16 @@ public class FailP4RequestErrorHandler
         return null;
     }
 
+    @Override
+    protected boolean isRetryableError(@NotNull Exception e) {
+        return false;
+    }
+
+    @Override
+    protected int getMaxRetryCount() {
+        return 0;
+    }
+
 
     @Override
     public void handleOnDisconnectError(@Nonnull ConnectionException e) {