+++ idea-p4server/impl/src/test/java/net/groboclown/p4/server/impl/connection/TestableP4RequestErrorHandler.java
@@ -60,6 +60,11 @@ public class TestableP4RequestErrorHandler
         return super.handleError(info, e);
     }
 
+    @Override
+    protected int getMaxRetryCount() {
+        return 0;
+    }
+
 
     @Override
     public void handleOnDisconnectError(@Nonnull ConnectionException e) {